% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cedup_avaliacoes_notas.R
\name{get_cedup_avaliacoes_notas}
\alias{get_cedup_avaliacoes_notas}
\title{Funcao para avaliacoes}
\usage{
get_cedup_avaliacoes_notas(.file_pdf)
}
\arguments{
\item{.file_pdf}{arquivo contendo horarios dos professores em pdf}
}
\description{
funcao para recuperar as avaliacoes por tipo e data a partir do diario de classe do professor
}
