% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cedup_cadastro_alunos.R
\name{get_cedup_cadastro_alunos}
\alias{get_cedup_cadastro_alunos}
\title{Função extrair dados de cadastro}
\usage{
get_cedup_cadastro_alunos(.files_xlsx)
}
\arguments{
\item{.files_xlsx}{arquivo contendo cadastro dos alunos em xlsx}
}
\description{
função para carregar cadastro dos alunos de xlsx para o R
}
