% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cedup_faltas.R
\name{get_cedup_faltas}
\alias{get_cedup_faltas}
\title{Função para extraçao de dados}
\usage{
get_cedup_faltas(.file_pdf)
}
\arguments{
\item{.file_pdf}{arquivo contendo diários de classes em pdf}
}
\description{
função para transformar pdf dos diarios em tabelas
}
