% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cedup_faltas_dia.R
\name{get_cedup_faltas_dia}
\alias{get_cedup_faltas_dia}
\title{Funcao para extracao de dados}
\usage{
get_cedup_faltas_dia(.file_pdf)
}
\arguments{
\item{.file_pdf}{arquivo contendo diarios de classes em pdf}
}
\description{
funcao para transformar pdf dos diarios em tabelas
}
