% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_horario.R
\name{get_horario}
\alias{get_horario}
\title{Função para conversão de horários}
\usage{
get_horario(.file_xltx)
}
\arguments{
\item{.file_xltx}{arquivo contendo horários dos professores em xltx}
}
\description{
função para transformar xlsx dos horários em tabelas
}
