% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cedup_cadastro_alunos.R
\name{save_cedup_cadastro_alunos}
\alias{save_cedup_cadastro_alunos}
\title{Função para salvar cadastros na nuvem}
\usage{
save_cedup_cadastro_alunos(.df_alunos_xlsx, .ss_id)
}
\arguments{
\item{.df_alunos_xlsx}{arquivo contendo cadastro dos alunos em xlsx}

\item{.ss_id}{do arquivo dos horário do google planilhas}
}
\description{
função para carregar cadastro dos alunos de xlsx para o R
}
