% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ead_drive.R
\name{save_ead_drive}
\alias{save_ead_drive}
\title{Função para conversão de horários EAD}
\usage{
save_ead_drive(.file_xltx, .ss_id)
}
\arguments{
\item{.file_xltx}{arquivo contendo horários dos professores em xltx}

\item{.ss_id}{do arquivo dos horário do google planilhas}
}
\description{
função para transformar xlsx dos horários em tabelas do EAD
}
