% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_horario_drive.R
\name{save_horario_drive}
\alias{save_horario_drive}
\title{Função para conversão de horários das aulas}
\usage{
save_horario_drive(.df, .turma, .ss_id)
}
\arguments{
\item{.df}{data.frame contendo horários dos professores em xltx}

\item{.turma}{vetor contendo turmas que serão salvas}

\item{.ss_id}{do arquivo dos horário do google planilhas}
}
\description{
função para salvar horários das aulas no google drive
}
